feat: add Theta Testnet support and larger icon
Support for Theta Testnet has been added, replacing the previously available blockchains. The icon size for Theta has been increased to improve its visibility. The code changes include the replacement of the filecoin calibration and hardhat chains with thetaTestnet, modification of the ThetaFilled icon size, and addition and adjustment of constants related to thetaTestnet. This allows for expansion of possible networks within the app and improves user interaction with the network icon.